refactor: split test_openmeteo_integration.py into focused modules
Split 541-line test_openmeteo_integration.py into 5 focused modules under 300 lines each:

Before:
- test_openmeteo_integration.py: 541 lines

After:
- tests/openmeteo_integration/conftest.py: 76 lines (shared fixtures)
- tests/openmeteo_integration/test_basic_integration.py: 156 lines
- tests/openmeteo_integration/test_fallback_mechanisms.py: 81 lines
- tests/openmeteo_integration/test_concurrency.py: 48 lines
- tests/openmeteo_integration/test_temperature_handling.py: 228 lines
- tests/openmeteo_integration/__init__.py: 1 line
- Total: 590 lines (49 lines added for better organization)

Key improvements:
- All files under 300-line limit for better maintainability
- Logical functional separation by integration aspects
- Shared fixtures in conftest.py reduce code duplication
- All 18 tests passing with no functionality lost
- Consistent naming convention following established patterns
- Proper test isolation with focused modules

Functional organization:
- Basic Integration: Core service selection and data retrieval
- Fallback Mechanisms: Service fallback behavior and error handling
- Temperature Handling: Unit preferences, conversions, and UI integration
- Concurrency: Thread safety and concurrent request handling

Maintains backward compatibility and test coverage while significantly
improving code organization and maintainability.

Task: Continue test file refactoring following successful weather_service pattern
