+++ tests/openmeteo_integration/__init__.py
@@ -1,0 +1,1 @@
+"""OpenMeteo integration tests package."""
